CP Registration
Control Point Registration between two images. This code is used to register the Koshy reference atlas images to the Alan atlas images. It saves (as .mat MatLab variable files) both tforms (affine and piecewise) as well as both sets of control points used, and both the intermediate and the final image.